Update entrypoint to take in an input argument
After discussion, our method for cumulus init and cumulus
start is:

For init, use docker-compose run [service] INIT, for every service.
This will start the container and run the default entrypoint with
INIT as an input argument

For start, simply use docker-compose up

Note that this means that "start" will create entirely new containers
(since docker-compose up cannot use the same container that was
used by docker-compose run). So, no state can be stored in the
containers after "init" (which is probably a good paradigm to follow)